Quito de HomeScreen al componente image y lo llevo a uno nuevo con su types.jsx y styles.tsx
