+++ src/components/Image/styles.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+const StyledImageResponsive = styled.img`
+    display: block;
+    max-width: 100%;
+`;
+
+export { StyledImageResponsive };
