+++ src/components/Image/types.ts
@@ -1,0 +1,4 @@
+export interface IImage {
+    alt: string;
+    src: string;
+}
